feat(runtime): slice-4 extract NostrSession runtime composition with PWA adapter
