+++ taskify-runtime-nostr/src/index.ts
@@ -4,3 +4,4 @@ export * from "./EventCache.js";
 export * from "./PublishCoordinator.js";
 export * from "./CursorStore.js";
 export * from "./SubscriptionManager.js";
+export * from "./RuntimeNostrSession.js";
